requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
pydantic==2.9.2
email-validator==2.2.0
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
pytest==8.3.3
httpx==0.27.2
